Var Selector
This program returns a an excel file with cosine similarity analysis between a set of variables description. This is because some Data Bases doesn't have clear specifications about their variables and they could contain different names for a same variable in all the files of the Data Base. With this program I aim to avoid to test variable per variable.

 The excel output is divided into three sheets: 

- presence: Show a matrix of the .csv files name as rows and variables name as columns. If the variable is in the file, the value is 1. Else, the value is 0.

- similarity: The cosine similarity matrix result of the variables.

- pairs: At each similarity value from 0.01 to 0.99, shows the pairs with that similarity proportion.

To run the program you need a directory with only the .csv files. Each of them should have the name with the format: "survey name+_+survey division+_+year+month+.csv"
all in lower case, example: enemdu_persona_2007_06.csv. If the survey doesn't have a division, omit it. If the survey has only an annual frequency, uses 00. If the survey has a monthly or quarterly, put the month with two digit ( 01, 02, 03, ..., 10, 11, 12).

 In the same directory where you are running this code, you'll need de description file with the name "survey name+_var_names_description.csv", ex: enemdu+_var_names_description.csv. The description file has a column with the variables names and other with their descriptions. The first row should be "variable"+separator+"description", that is, the first row has the columns name.